chore: migrate zlib to use BUILD.bazel from BCR

+++ source/extensions/compression/gzip/compressor/zlib_compressor_impl.cc
@@ -31,8 +31,10 @@ ZlibCompressorImpl::ZlibCompressorImpl(uint64_t chunk_size)
 void ZlibCompressorImpl::init(CompressionLevel comp_level, CompressionStrategy comp_strategy,
                               int64_t window_bits, uint64_t memory_level = 8) {
   ASSERT(initialized_ == false);
-  const int result = deflateInit2(zstream_ptr_.get(), static_cast<int64_t>(comp_level), Z_DEFLATED,
-                                  window_bits, memory_level, static_cast<uint64_t>(comp_strategy));
+  const int result =
+      deflateInit2(zstream_ptr_.get(), static_cast<int>(comp_level), Z_DEFLATED,
+                   static_cast<int>(window_bits), static_cast<int>(memory_level),
+                   static_cast<int>(comp_strategy));
   RELEASE_ASSERT(result >= 0, "");
   initialized_ = true;
 }
